Moves all services to E5R.Tools.Bit.Engine
