+++ src/E5R.Tools.Bit/E5R.Tools.Bit.csproj
@@ -6,9 +6,6 @@
     <AssemblyName>bit</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.0.0" />
-  </ItemGroup>
-  <ItemGroup>
     <ProjectReference Include="..\E5R.Sdk.Bit\E5R.Sdk.Bit.csproj" />
     <ProjectReference Include="..\E5R.Tools.Bit.Engine\E5R.Tools.Bit.Engine.csproj" />
   </ItemGroup>
