+++ test/E5R.Tools.Bit.Test/ProgramTests.cs
@@ -8,7 +8,6 @@ using Moq;
 
 namespace E5R.Tools.Bit.Test
 {
-    using Engine;
     using Engine.DI;
     using Sdk.Bit.Services.Abstractions;
     using _ = E5R.Test.Commons.TraitConstants;
@@ -17,18 +16,11 @@ namespace E5R.Tools.Bit.Test
     [Trait(_.COMPONENT, nameof(Program))]
     public class ProgramTests
     {
-        [Fact(DisplayName = "Constructor does not accept null for [env, container] parameters.")]
+        [Fact(DisplayName = "Constructor does not accept null for [container] parameter.")]
         public void Constructor_Does_Not_Accept_Null()
         {
-            var envMock = new Mock<IBitEnvironment>().Object;
-            var servicesMock = new Mock<IServiceCollection>().Object;
-            var containerMock = new Mock<DependencyInjectionContainer>(servicesMock).Object;
-
-            var errorEnv = Assert.Throws<ArgumentNullException>(() => new Program(null, containerMock));
-            Assert.Equal("env", errorEnv.ParamName);
-
-            var errorContainer = Assert.Throws<ArgumentNullException>(() => new Program(envMock, null));
-            Assert.Equal("container", errorContainer.ParamName);
+            var error = Assert.Throws<ArgumentNullException>(() => new Program(null));
+            Assert.Equal("container", error.ParamName);
         }
     }
 }
